requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-cron==0.6.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-mptt==0.15.0
django-redis==5.4.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
jdatetime==4.1.1
kavenegar==1.1.2
lxml==4.9.3
Pillow==10.1.0
platformdirs==3.11.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pypng==0.20220715.0
pyshorteners==1.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
zeep==4.2.1
